% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eval_lmm}
\alias{eval_lmm}
\title{Calculates mean at midpoint for each individual's trajectory}
\usage{
eval_lmm(df, id_name, time_name, individual_coeffs)
}
\arguments{
\item{df}{dataframe}

\item{id_name}{column name that contains the subject id value, e.g. "ID"}

\item{time_name}{column name that contains the time value, e.g. "Time"}

\item{individual_coeffs}{dataframe output from get_individ_lm()}

\item{value_name}{column name that contains the biomarker value, e.g. "Value"}
}
\value{
eval_at_midpoint dataframe containing mean at midpoint of individual trajectory
}
\description{
Calculates mean at midpoint for each individual's trajectory
}
