% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_polynomial}
\alias{fit_polynomial}
\title{Fit a non-negative polynomial - helper function}
\usage{
fit_polynomial(x, y, degree = 3)
}
\arguments{
\item{x}{time variable}

\item{y}{biomarker variable}

\item{degree}{degree of polynomial fit. Budgeon et al suggests 3 is sufficient, but 5 is also good.}
}
\value{
fit polynomial model
}
\description{
Fit a non-negative polynomial - helper function
}
