% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_synth_data}
\alias{generate_synth_data}
\title{Generates synthetic data}
\usage{
generate_synth_data(n, T, l)
}
\arguments{
\item{n}{Number of individuals}

\item{T}{total length}

\item{l}{length of time between observations}
}
\value{
all_data synthetic dataset with n participants with unique ID's, observations at an interval of Time l, and Marker_level values based on distribution parameters
}
\description{
Generates synthetic data
}
