% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_horizontal_adjustment}
\alias{get_horizontal_adjustment}
\title{Shifts model output so that "0" is when the biomarker becomes positive
Budgeon et al starts all models at time 0, this shift is done to make the model output consistent with SILA (Betthauser et al, 2022)}
\usage{
get_horizontal_adjustment(df, PET_pos_threshold)
}
\arguments{
\item{df}{dataframe containing the IDs, biomarker values and timepoints}

\item{PET_pos_threshold}{threshold for biomarker positivity}
}
\value{
hor_adj single number used to shift model output time
}
\description{
Shifts model output so that "0" is when the biomarker becomes positive
Budgeon et al starts all models at time 0, this shift is done to make the model output consistent with SILA (Betthauser et al, 2022)
}
