% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_individ_lm}
\alias{get_individ_lm}
\title{Performs ID-by-ID linear regression}
\usage{
get_individ_lm(df, value_name, time_name, id_name)
}
\arguments{
\item{df}{dataframe}

\item{value_name}{column name that contains the biomarker value, e.g. "Value"}

\item{time_name}{column name that contains the time value, e.g. "Time"}

\item{id_name}{column name that contains the subject id value, e.g. "ID"}
}
\value{
individual_coeffs dataframe containing slopes and intercepts on an ID-by-ID basis
}
\description{
Performs ID-by-ID linear regression
}
