% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_integration_estimates}
\alias{get_integration_estimates}
\title{Integrates per step 3 in Budgeon et al process}
\usage{
get_integration_estimates(eval_at_midpoint, id_name, degree = 3)
}
\arguments{
\item{eval_at_midpoint}{dataframe containing mean at midpoint of individual trajectory}

\item{id_name}{column name that contains the subject id value, e.g. "ID"}
}
\value{
dataframe with tau and mu_tau (step 4)
}
\description{
Integrates per step 3 in Budgeon et al process
}
