% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_Time_to_Positivity}
\alias{get_Time_to_Positivity}
\title{Wrapper function for entire process}
\usage{
get_Time_to_Positivity(
  df,
  id_name,
  time_name,
  value_name,
  PET_pos_threshold,
  degree = 3
)
}
\arguments{
\item{df}{dataframe}

\item{id_name}{column name that contains the subject id value, e.g. "ID"}

\item{time_name}{column name that contains the time value, e.g. "Time"}

\item{value_name}{column name that contains the biomarker value, e.g. "Value"}

\item{PET_pos_threshold}{numeric threshold for biomarker positivity}

\item{degree}{degree of polynomial fit}
}
\value{
df_tau dataframe containing ID, predicted value based on model fit, time to positivity estimate
}
\description{
Wrapper function for entire process
}
