% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{poly_ft}
\alias{poly_ft}
\title{Fit a non-negative polynomial}
\usage{
poly_ft(eval_at_midpoint, id_name, degree = 3)
}
\arguments{
\item{eval_at_midpoint}{dataframe containing mean at midpoint of individual trajectory}

\item{id_name}{column name that contains the subject id value, e.g. "ID"}
}
\value{
to_return dataframe with polynomial fits corresponding with IDs
}
\description{
Fit a non-negative polynomial
}
